+++ packages/features/src/Domain/Component/EditorIdentifier.ts
@@ -1,5 +1,0 @@
-import { FeatureIdentifier } from './../Feature/FeatureIdentifier'
-
-type ThirdPartyIdentifier = string
-
-export type EditorIdentifier = FeatureIdentifier | ThirdPartyIdentifier