+++ packages/services/src/Domain/Feature/FeaturesClientInterface.ts
@@ -1,12 +1,15 @@
-import { FeatureIdentifier } from '@standardnotes/features'
 import { ComponentInterface, DecryptedItemInterface } from '@standardnotes/models'
 
 import { FeatureStatus } from './FeatureStatus'
 import { SetOfflineFeaturesFunctionResponse } from './SetOfflineFeaturesFunctionResponse'
+import { NativeFeatureIdentifier } from '@standardnotes/features'
+import { Uuid } from '@standardnotes/domain-core'
 
 export interface FeaturesClientInterface {
-  initializeFromDisk(): void
-  getFeatureStatus(featureId: FeatureIdentifier, options?: { inContextOfItem?: DecryptedItemInterface }): FeatureStatus
+  getFeatureStatus(
+    featureId: NativeFeatureIdentifier | Uuid,
+    options?: { inContextOfItem?: DecryptedItemInterface },
+  ): FeatureStatus
   hasMinimumRole(role: string): boolean
 
   hasFirstPartyOfflineSubscription(): boolean
@@ -16,13 +19,13 @@ export interface FeaturesClientInterface {
 
   isThirdPartyFeature(identifier: string): boolean
 
-  toggleExperimentalFeature(identifier: FeatureIdentifier): void
-  getExperimentalFeatures(): FeatureIdentifier[]
-  getEnabledExperimentalFeatures(): FeatureIdentifier[]
-  enableExperimentalFeature(identifier: FeatureIdentifier): void
-  disableExperimentalFeature(identifier: FeatureIdentifier): void
-  isExperimentalFeatureEnabled(identifier: FeatureIdentifier): boolean
-  isExperimentalFeature(identifier: FeatureIdentifier): boolean
+  toggleExperimentalFeature(identifier: string): void
+  getExperimentalFeatures(): string[]
+  getEnabledExperimentalFeatures(): string[]
+  enableExperimentalFeature(identifier: string): void
+  disableExperimentalFeature(identifier: string): void
+  isExperimentalFeatureEnabled(identifier: string): boolean
+  isExperimentalFeature(identifier: string): boolean
 
   downloadRemoteThirdPartyFeature(urlOrCode: string): Promise<ComponentInterface | undefined>
 }