+++ packages/api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.32](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.31...@standardnotes/api@1.26.32) (2023-07-26)
+
+**Note:** Version bump only for package @standardnotes/api
+
 ## [1.26.31](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.30...@standardnotes/api@1.26.31) (2023-07-26)
 
 **Note:** Version bump only for package @standardnotes/api