+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.56.27](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.26...@standardnotes/mobile@3.56.27) (2023-07-26)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.26](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.25...@standardnotes/mobile@3.56.26) (2023-07-26)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.56.25](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.24...@standardnotes/mobile@3.56.25) (2023-07-26)
 
 **Note:** Version bump only for package @standardnotes/mobile