+++ packages/snjs/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.202.16](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.15...@standardnotes/snjs@2.202.16) (2023-07-26)
+
+### Bug Fixes
+
+* Fixes issue where lock screen would not use previously active theme ([#2372](https://github.com/standardnotes/app/issues/2372)) ([d268c02](https://github.com/standardnotes/app/commit/d268c02ab31beb5e2fd9e6547610f9a4dd61bed4))
+
+## [2.202.15](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.14...@standardnotes/snjs@2.202.15) (2023-07-26)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
 ## [2.202.14](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.13...@standardnotes/snjs@2.202.14) (2023-07-26)
 
 **Note:** Version bump only for package @standardnotes/snjs